# Vectorize audio downsampling.