Benchmarks:

```
name                                             old time/op             new time/op             delta
BM_Resample<false, 300, 22050, 8000>             36.9ms ± 1%             30.4ms ± 1%  -17.82%         (p=0.000 n=10+9)
BM_Resample<false, 300, 15000, 8000>             24.7ms ± 0%             23.4ms ± 1%   -5.27%          (p=0.000 n=9+9)
BM_Resample<false, 100, 4000, 400>               6.27ms ± 0%             3.88ms ± 1%  -38.10%         (p=0.000 n=9+10)
BM_Resample<true, 300, 22050, 8000>               124ms ± 1%               95ms ± 1%  -22.90%        (p=0.000 n=10+10)
BM_Resample<true, 300, 15000, 8000>              84.8ms ± 1%             75.2ms ± 1%  -11.26%         (p=0.000 n=9+10)
BM_Resample<true, 100, 4000, 400>                22.1ms ± 1%             19.2ms ± 2%  -13.31%         (p=0.000 n=9+10)
```